requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
streamlit==1.44.1
python-docx==1.1.2
reportlab==4.4.1
pydantic==2.11.3
python-multipart==0.0.20
httpx==0.28.1
